Provide some basic testing models
This commit adds bare models for dormitory, room and user. They are basicly for demonstration purposes there.

You can test it with a python repl:
>>> from pycroft import model
>>> model.create_db_model()
>>> model.session.session.flush()

If everything is successfull installed this will create a database file named test_db.sqlite in the toplevel dir. You now can dump the schema with the sqlite-commandlinetool:
$ sqlite3 test_db.sqlite .schema

This should return the three table-definitions.

fixes #136 @1h