Fix ordered backrefs
The order_by backref argument expects either a callable (which then gets
called) or a string with a column expression. order_by=id is unfortunately
not interpretede as an ordering on the `id' column, but the builtin id(),
which then crashes since it expects an argument, but is called without one.

Signed-off-by: Maximilian Marx <mmarx@wh2.tu-dresden.de>