Add a new BlueprintAccess class and extend BlueprintNavigation
This commit adds a new class called BlueprintAccess that is a per-Blueprint
access registry. It can be uses to restrict access of single view functions to
users with special given permissions or general to logged in users.
It also integrates the new class with the BlueprintNavigation to show only
the navigation menu entries to which the user has access. With this I also
moded the BlueprintNavigation to a new module web.blueprint.navigation.

refs #199