Revert "optimized code style with pycharm"
This reverts commit 88534b5be96dcc642cdf4a9bf3ebe2f27a12c86d.

Please do NEVER "optimize such a huge amount of code for
style reasons. It makes the whole code un-blameable, it
makes it very difficult to revert changes, its a no1 cause
for conflicts and in a lot of cases - special for minified js
and css files it doesn't make any sense.

And please: the html files should not follow the 80-char-line
pep8-doctrin (the one thing of pep8 I really doesn't
understand because everyone have a form of widescreen
display today - so vertical space is a lot more expensive
than horizontal).

Conflicts:

	web/blueprints/user/__init__.py
	web/templates/user/add_membership.html
	web/templates/user/user_show.html