improve subnavigation handling
This fixes the huge space between the sections with proper calculations
of scroll corrections (I hope it will look equal in all browsers) and wraps
it to a jQuery plugin for easier reuse in other views.

It also fixes the corner case when the document end is reached. Then it
must not correct the scrolling position.

Note that subnav-fixed will be set only by the plugin. If you set in manually
within the html source the offset calculations will get wrong initial values.