fix a bug in membership disabling
If we disable a future membership we get an error from the validator because
the end_date is before the start date. I fixed this using the start_date as end_date
if now is before start_date. This can cause an disabled membership to be active but
only for a minimal time interval.

The bug was found while writing unit tests. So: write tests!