refactor the get_free_ip helper and add a new select_subnet_for_ip helper
This changes the get_free_ip helper method to work properly. It first gets all
assigned addresses from the subnets NetDevices. Then it checks if there are
free ips possible. If there are possibly free ips it iterates all the host-ips of the
network and select the first free one.

The select_subnet_for_ip can be used to select the subnet that contains an
ip from a list of subnets