fix circular dependency issue, caused by the switch-management_ip
The management_ip of the switch needs a network-device. the network-device needs
an host - the switch. but the switch needs the management_ip. So we have a cycle
that cannot be simply inserted in a database.

The sqlalchemy-way to fix this, is to use a ForeignKey with use_alter=True and a
relationship with post_update=True. The details are in the docs at [1].

This has one mayjor disadvantage: You have to set the foreign key nullable. The
reason is simple: The insert without te fk will fail with NOT NULL set in the database.
So we have to assert that the management_ip was set. This check is done in the
pre_{insert,update}-event-hook. The hook also checks that the ip is set to at
least one network_device of the switch.

[1] http://docs.sqlalchemy.org/en/latest/orm/relationships.html#rows-that-point-to-themselves-mutually-dependent-rows