Changed the model for NetDevice and it's derived classes
Changed the relationship of the derived classes from the NetDevice
class, in the way, that the id for the relationship can stay in the
base class and the realationship in the derived class.
The version via primary_join didn't work, because of the intern
structure of sqlalchemy. So I renamed the realtionship of the derived
classes to host, so the fit to the column host_id in the base class.