Changed library functions of the user to fit new model
Fixed some errors in the move_in and move_to functions of the user
library. They now use UserNetDevices in stat of NetDevices, because
they are the abstract class. Furthermore some other parts were changed,
because the name for the relationship from user to netdevice changed.
In addition the relationship changed changed from "many to one" to "one to one"
witch makes the traffic-accounting and other thinks easier.