Added cascades for dns records to the model
This commit adds sql cascades to the model for the host aliases as well
as tests for these cascades. A special position is owned by the ARecords
and the AAAARecords, because I didn't want a backref from Ip to these
record types. So every time an Ip is deleted a method is trickered which
checks whether there are still some records which point to the specific
Ip and throws an exception if this is the case.