Removal of logic from model and templates -- hostalias
This commits removes some logic from the model and a template, which was
used for the host aliases.
The function of the model which returned a human readable name of the
record was replaced by a jinja filter.
Additionally, the decision whether a record could be deleted or edited
was moved from the template to another jinja filter.

refs #391