refactor transaction date parsing
Now we use strptime to parse the date. This has the
benefit that we get an erly exception if the date is
not correct formatted.

This also uses relativedelta to guess the correct year
for the date because the value to parse have omitted
this information. We guess by assigning the current year
and date it on year back if the resulting date is in
the future.

references #464
fixes #465
fixes #466