Add grace period and proper fee fields to semseter
The semester was missing a grace period, now specified with
premature_begin_date and belated_end_date, in which a member
can register prematurely or cancel belatedly without being charged
for an extra semester. Plain dates are used instead of datetime,
which makes more sense.

Second, fields for the overdue fine and the reduced membership fee
were missing. For consistency reasons, semester fee is renamed to
regular membership fee.