Add static form fields
Field may be wrapped in a call to the static field function, which
will cause the field to be rendered as a Bootstrap static form
control. Do not confuse static fields with read only fields.
Both can not be edited by the user, but are rendered differently.
Read only fields suggest to the user that he can not edit the field
right now, because he lacks some permissions for example. Static
fields on the other hand are really static, they will never be
editable.