Make use of Bootstrap widgets explicit
Replace the widget monkey patching with explicit imports. For all
WTForm core fields and the sqlalchemy fields, field classes with
the same have been provided. The only thing that needs to be done
in order to use the field with a Bootstrap widget is replace the
package name in the import statements.

This increases flexibility as it may sometimes be necessary to use
the non-decorated version of a field and it doesn't involve behind
the scenes replacement of internal stuff, which is a bit evil.