Revert "Rename template_check to template_test"
We have the team policy to not prefix code parts with "test" because
this confises the nose test discovery process. We need to give nose all
out code because otherwise it does not find modules that are not tested
for the coverage report.

Eventually we find a better solution at some point, but at the moment
this is the only way to get working tests and overall coverage reports.

I created a marker bug #526 to track this problem.

This reverts commit c6142b268f8cfb2182ad9c92b09fe776e3ba2b24.