Move with_transaction decorator into the session file
The with_transaction decorator was previously located in the lib/all.py
file. As this wrapper is strongly coupled with the session which is
used, the function was moved to this file.

Furthermore the log of currently running transactions is saved inside
the session and not in a global variable. This enables us to have
multiple session run in parallel and don't become confused with the
transactions which are on the fly.