Implement time-based membership queries
Add the ability to query the
* memberships
* groups
* properties
of a user within a given interval instead of just without time and
active right now.

It was not possible to implement the active_property_groups and
active_traffic_groups methods of the User class as hybrid_method.
If anybody has an idea how to accomplish this, it would be very
welcome.