Implement fees
This commit implements fees. Each fee is implemented as class that
has the specific logic for a given fee. Fee classes compute the
credit transactions if the users owe us money. The idempotent
post_fees function uses those Fee classes to compute the transactions
that should have been posted and compares it to the transactions that
have been posted. Missing transactions or adjustments for wrongly
posted transactions are automatically posted.