Use object_session to obtain a session in model instances
Hybrid properties and methods often retrieve additional data from
the database. Instead of using the instance's session, they used
the global session obtained via import. This is bad for several
reasons:
* First of all, the idea of having a global session, you can freely
import from anywhere you are, was a very bad idea on its own and we
should get rid of it.
* If different sessions are active and the global session is not
the one the object is actually associated with, state tracked in the
associated session is not respected.