Validate room input before calling lib.user.move()
Previously, leaving the same room in the UserMoveForm activated an
assert which lead to printing the complete debug ui.
Now, the given data is checked before calling lib.user.move(), flashing
a normal error message.

I had to reposition the part where the form data is filled with the
defaults, because this has to be done too, if the same room is given
(decided by bool refill_form_data) -- else, the fields would be left empty.