Add formatter & column-structure for bs-tables
Previously, a table column (cell) could only be transmitted as a string.
This made displaying links extremely unelegant, which is why I did the
following:

1) Use a list of dicts for the display_columns parameter in the
html-macro which creates the table structure. An example for this can be
found in the switch_show.html (see diff).
In this structure, one can assign different `formatters` to the column
which are implemented as js-fncs in custom\/js/table.js.

2) Implement a linkFormatter() which processes data of the assigned
column (rather cells of these) as json instead of a string. This json
must have the properties 'href' and 'title'.

3) Use the linkFormatter() for the `Raum`-column in switch_show().