Add vagrant docker support
- the provision script has been simplified and is only used for
  VirtualBox
- PostgreSQL is started through supervisord with a startup script
- the PostgreSQL startup script recreates a cluster and the databases
  in a tmpfs if necessary
- PostgreSQL is tuned to run tests by disabling durability features
- the docker provider is set in the PyCharm Vagrant settings

Please note:
After adding a remote interpreter, PyCharm will upload the PyCharm
helpers to the vagrant user's home directory. As docker containers
are ephemeral those helpers will be gone if the container is restarted.
Therefore change the PyCharm helpers path to
/pycroft/.pycharm_helpers
after adding the remote interpreter. You can only do this in the
edit dialog, not in the add dialog unfortunately. This will place
the PyCharm helpers in the project root, which are not part of the
container and thus persist.