Add vagrant user UID/GID customization support for docker
Ensuring that the vagrant user inside Docker has the same UID/GID
as the developer running vagrant is very challenging. Docker tries
very hard to enforce host-independent builds and rightly so. But
in the case of Vagrant this is very unfortunate.

The solution implemented here is kind of hack. The 'defaults' file
inside the vagrant subdirectory is not tracked via git, but needed
to start the docker container. Developers should specify the UID
and GID of the user who owns the files in the source try using the
two variables VAGRANT_UID and VAGRANT_GID.

An alternative would be to use NFS and its UID/GID mangling
capabilities, but would require NFS support on the developer
machine.