Enhance a user's flags (prev. “status”)
I reviewed the former `determine_user_status` and thought that the
things returned are nothing but some kind of flags with informational
character which are only(!) used in the `user_show_basedata.html`.
Because (in this context) it is rather useless to assign each of these
flags some kind of identifier, as i.e. `traffic_exceeded`, I modified
the function in order to immediately return the actual GUI
_description_, i.e. “Traffic Übrig”, which removes the need to insert
them statically in the html template -- which now dynamically _loops_
over every flag to display it.