Enhance CSV import
The CSV/MT940 import was not working good enough. The entries in
the database are now diffed with the new entries. This allows users
to upload CSV files which contain already imported entries.

Several error checking mechanisms have been implemented. Most
importantly the user has to provide an expected balance that must
be present after the import.

Fixes #562, #447