Overhaul DNS model
This commits does some big overhaul of the DNS model. The model was
not quite right. Most records were associated with hosts and some
records such as CNAME were associated with other records. This is
not how DNS really works and therefore leads to problems.

Fundamentally DNS associates hierarchical names with data, where
the data can be IP addresses (A, AAAA), other names in the DNS
(CNAME, MX, NS, SOA, PTR, SRV) or arbitrary data (TXT, SRV).
A new class of objects DNSName is introduced which belong to a DNSZone.
Records always have a DNSName and depending on their type are associated
with other objects. AddressRecords (a type that combines A and AAAA records)
for example are associated with IPs, CNAMERecords are associated with
a second DNSName etc.

In addition the user-defined types SQLAlchemy types IPAddress, IPNetwork
and MACAddress are introduced, which automatically use specialized
Postgres types if possible and fallback to string in other dialects and
in the case of IPAddress and IPNetwork use objects from ipaddr package
on the Python side.