Isolate database tests with transactions
All tests are run inside an external transaction with a SAVEPOINT set
after the fixtures have been loaded. The transaction is rolled back
right before the fixtures are teared down, so that fixtures can be
deleted properly without interfering with any left behind objects
from the tests.
The transaction is external to the ORM session, so that it may safely
issue commits and rollbacks. See [1] for details. The underlying database
must support savepoints which both SQLite and of course Postgres do.
Furthermore the use of package-level setup and teardown functions,
supported by nose speeds up tests considerably, especially on Postgres,
as tables are only created once during test execution.

[1] http://docs.sqlalchemy.org/en/rel_0_9/orm/session_transaction.html#joining-a-session-into-an-external-transaction-such-as-for-test-suites