Remove .pycharm-helpers from .gitignore
Moving the remote pycharm-helpers folder into the project directory
and adding it to the .gitignore file was based on a misunderstanding
of how the vagrant docker provider works.

It was assumed that it would create a new docker container each time
the vagrant machine is started, this is however not the case. The
same container is launched every time unless a recreation is explicitly
requested with docker reload.