Remove lowercase {crypt} special case.
The lowercase {crypt} LDAP password hash scheme is actually a wrong
AG DSN implementation in python-takeover. OpenLDAP always creates
uppercase hash type identifiers and does not distinguish lowercase
and uppercase identifiers and treats them equally.

Passlib on the other does not accept lowercase identifiers and will
raise an ValueError. Instead of handling this in the Pycroft, our
current LDAP records have been fixed.

BTW: The correct way to address this issue in the code would have
been to convert the identifier to uppercase instead of special-casing
and reimplementing a certain crypt algorithm, which is already
supported by passlib.