Make ldap_sha512_crypt the default password hash
The crypt algorithm with the ID 6 is the best available and
supported password hash algorithm in our infrastructure. It is far
superior to SSHA, because SSHA is just SHA-1 with a single round,
whereas crypt algorithm 6 is SHA-512 with a large and configurable
number of rounds, passlib currently uses 100000 rounds.