Rename Journal to BankAccount
Presumably the name journal was chosen to distinguish bank accounts from the
accounts in Pycroft's double-entry bookkeeping system. The problem is, journal
is the wrong name. In accounting, a journal is the record of all financial
transactions in a bookkeeping system ordered by date. The German equivalent is
the “Grundbuch”. It's the chronological view of all business activity, whereas
the accounts in the “Book of Accounts”/“Hauptbuch” present a view sorted by
topic or type of business activity.

This commit renames Journal to BankAccount and JournalEntry to
BankAccountActivity