Fix docker-compose files, use tmpfs for postgres
- Use tmpfs for the pg container

- Remove `links` section from bare, add docker-compose-travis.yml

The latter is necessary because docker-compose doesn't allow extending
services with predefined links (in this case, `web` in
`docker-compose.bare`).  Since we don't want a volume on the
docker-compose file travis will use, but still need the link, we have to
introduce a separate file only to define the link.

However, this is still better than using two explicit, redundant files
which may get out of sync.