Reconstruct customly small membership intervals
Sometimes, instead of the regular membership fee, a smaller one has been
booked by the head of finance, to compensate for the smaller interval of
membership.  In most cases, this can be used to reconstruct the
membership intervals in higher precisions (instead of throwing an error
due to fee amount inconsistencies).