Fix tests: Don't set an explicit UnixAccount.uid
This fixes the tests because the FixtureDataTestBase doesn't reset
sequences – in this case, _uid_seq ran higher and higher across
tests.v, eventually reaching above 1024, which caused the collision
with the dataset setting an explicit uid.

Anyhow, testing that setting an explicit uid is possible is just no
use case and should not be done except for special purposes, such as
the initial importer.  On these occasions, one should fix the
sequences when finished – as is done in the legacy importer.