Use canonical escaping for record values
This fixes a bug that if special chars according to RFC04515 are
present in an attribute's value, comparison fails against records
retrieved from an LDAP lookup, since said characters are escaped
there.

Now, additionally to the canonicalization of values to a list, string
or byte objects undergo escaping any time a Record object is
initialized.