Add rabbitmq and worker to tests, refactor testing setup
The docker-compose setup is now split up into multiple parts, allowing
for independent running of the testing instance, testing, and letting
travis test (which needs to be separate because it does not need
any volumes).

Also, it is extended by a `rabbitmq` broker being utilized by a dummy
celery worker which is / will be used in the tests.