Remove legacy parameters for user creation (UI and lib)
The parameters `moved_from_division` and `already_paid_semester_fee`,
and the respective form fields on the user creation page, are
factually obsolete and are removed from the frontend and the
`lib.user.move_in` function.

Note that this is in reference to #28, but does not fix it, because
the logic is still in the model.

Note that the `host_name` entry in the docstring was obsolete
beforehand.