Make Host.room non-nullable, adapt handling in importer
Any host should have a room.  The only reason hosts wouldn't have a
room in teh current dump would be if the room was ignored due to a
missing hp4108port.  This means that the room is not connected and
probably does not exist anymore, so instead, we add a nice warning and
drop the user's hosts.

Fix #127