Fix tests and Simplify `IntervalField` validation
The validation routine throw KeyErrors whervn too little tokens have
been supplied.  Also, an empty interval has been considered valid.

Because there is no python-like analogon to
`IntervalField.data` (timedelta is insufficient), it has been
simplified to just be the raw string – in the end, this is what's
being passed to SQLAlchemy as well.  All the syntax requirements are
handled in the validation routine.