Revert "Delegate a `Table` to a `View`"
This reverts commit 2c4033b0caf9e8aa3f129a27ef9ac9d2bbbc6258.

This was the wrong approach.  Instead of creating a `Table`, using the
more simple `TableClause` is sufficient as it does not require being
bound by the metadata.  Making the Table unvisitable inhibits it from
being usable in a select statement, as the compiler won't know what to
do with it, so the main goal was not even achieved.