Sanitize email in UserCreateForm
The the recent changes on email validation led to the failing of two
tests. If no email is passed to the UserCreateForm, an empty string is
set as the user's email instead of None. However, the changed email
validation function correctly detects that an empty string is not a
valid email address.