Adapt tests to new session.utcnow() implementation
The tests assumed that time returned by session.utcnow() changes between the
call to Membership.disable() and Membership.active(). That was the case with the
previous implementation of session.utcnow(), which used STATEMENT_TIMESTAMP()
(current date and time at the start of current statement) as its timestamp
provider. However, the new implementation uses CURRENT_TIMESTAMP (current date
and time at the start of current transaction) as its time provider. All tests
run inside their own nested transactions, but CURRENT_TIMESTAMP seems to return
the start timestamp of the outermost transaction. Consequently all calls to
CURRENT_TIMESTAMP return the same value, which leads to the failing of the
tests.