Fix search and sorting of unassigned bank account acctivity table
The table was configured to be paginated on server side. However, the
corresponding json endpoint incapable of handling the sort and search
arguments passed by bootstrap table. As the amount of unassigned bank
account activities should usually be within a manageable range, it is
easier to let bootstrap table handle all the pagination, filtering and
sorting.