Work around bug in LazyLoadSelectField's with multiple or chained conditions
Imagine we have three fields: building, level and room_number.
The field level depends on the field building.
The field room_number depends on the fields building and level.

If the value of building changes both level and room_number are
triggered to reload the options. As they are both triggered at the same
time, room_number uses the value of level from before the reload.
After level reloaded its options, a new value gets selected. However,
when programmatically changing the value of a select field, no on change
event is fired. Therefore room_number does not get informed that level's
value has changed.

The workaround is to manually fire an on change event, after the reload
of field's options is done. For fields with multiple conditions this can
lead to unecessary requests being sent.