Fix file-loader mappings in production
This undocumented option causes[1] the `file-loader` plugin to report
the files not as `export default $filename`, but as `module.exports =
$filename`. However, only the latter causes the `compilation` object to
get not only a `filename`, but also a `module` with the correct
`userRequest` value (e.g., `imgs/favicon.ico`).

This, in turn is used by the `webpack-manifest-plugin` to set up a
mapping in a hook:

```javascript
var moduleAsset = function (module, file) {
  console.log(module.userRequest, file);
  if (module.userRequest) {
    moduleAssets[file] = path.join(
      path.dirname(file),
      path.basename(module.userRequest)
    );
  }
};

// …
compiler.hooks.compilation.tap(pluginOptions, function (compilation) {
  compilation.hooks.moduleAsset.tap(pluginOptions, moduleAsset);
});
```

Later, `moduleAssets`, in turn, is used to set the `path` and the `name`
property accordingly. However, if the moduleAssets entry does not exist
for a file, the `name` value defaults to whatever `path` is.

Why this error only came up in a `prod` build I dare not say.

[1] https://stackoverflow.com/a/59075858/2443886