Turn `validate_unique_name` into proper wtforms validator
I'm not quite sure why this was done explicitly in the view functions.

In any situation, the form or field definition is where validation
logic belongs – executing the validator outside of the forms
validation results in `Form.validate_on_submit` not being able to
notice that there was invalid data.

Preparational refactoring for #428