Set `passwd_hash` for non-frontend tests
Unless `password` is set, we will directly set the `passwd_hash` to a
hash of the password `password`.

This cuts down the total execution time of our current test suite from
~300s to ~200s.

An even drastic difference can be observed when restricting to the
`lib` tests, since the slowest tests (`frontend/`) are unaffected by
this performance tuning.

The correct lesson to draw from that is that we should write less
frontend tests, and more `lib` tests or even pure unittests.

Weakly refs #15, #340