Turn assertion methods into functions
The only reason assertion methods have been subordinate to a class
deriving from `unittest.UnitTest` is that they had to utilize
`assertEquals` and related methods to provide proper context when the
test errored.  This is not necessary anymore since we use pytest as a
runner, which can provide such context for pure python `assert`
statements.

Thus, the reference to `self` is nowhere needed, and we can use flat
functions.

This is beneficial as it allows us to put assertions related to a
similar context into a central location (such as `user_assertions`),
making the tests more organized in the long term.

Also, it makes refactoring easier; in fact the motivation for this
change was to separate the user creation test from the `move_in` test.
Refactoring of tests is something we will have to do continuously if
we want to switch to pytest at some point.