Make some convenience relationships `viewonly=True`
These are transitive counterparts to more complicated relationships.
For instance, `TaskLogEntry.user` is just a shortcut for the (partial)
relation TaskLogEntry → (Task as UserTask) → User.

Obviously this overlaps with the actual foreign key relations, so it
triggers a warning.

Clearly it should be marked `viewonly=True`, as these relationships
exist only for convenience.

Refs #471