Add more status types to ADRs
“Postulated” means that the author just made that decision.
I think it is a useful verb to have because many decisions are made by
the people with expertise in the area they're working in, and does not
undergo a formal democratic decision process.

Since ADRs should be lightweight and should not draw people towards
unnecessary bueaurocratization, I think this is a good clarification
to make.