Add default celery transport options to avoid hanging test
See https://github.com/celery/celery/issues/4627#issuecomment-396907957:

If the broker is down, the timeouts we deal with don't quite grip
because these are timeouts waiting for a _response_ (i.e., what
happens after the `wait()` part).

However, if the broker is down, we don't even get to finish the
`apply_async()` call: we're dealing with _connection timeouts_.

Thhese timeouts can be configured by specifying a
retry policy in the `broker_transport_options`.  All of this is not
very well documented, or if it is, not that well discovered.

Frustratingly, `broker_connection_timeout`[1] seems to have no effect
whatsoever on the connection timeout strategy.

Now, the tests which fail the CI (by running in the timeout) properly
abort and show `ConnectionRefused`.

[1] https://docs.celeryproject.org/en/stable/userguide/configuration.html#std-setting-broker_connection_timeout