Fix SA usage warnings
Since the session is not guaranteed to be flushed before
execution (there is no fixture which is guaranteed to run last), and
`begin_nested` issues a flush, SA rightfully warns that there might be
objects lying around which are not persisted.

Usages have been adapted to force a flush before object creation.