Stop escaping of columns when a custom formatter is used
Since we use a custom formatter almost exclusively to deal with
complex JSON inputs at their source, we cannot make any use of the
builtin escaping, because it forcefully casts as strings.

In a change in bootstrap-table[1] from v1.19.0→v1.19.1, the escaping
forces a `.toString()` conversion instead of leaving values which are
JSON objects alone.  In this case, it technically just broke
functionality instead of adding security.

However, it is definitely necessary to escape user-controlled input in
our formatters, which we don't really seem to do at the moment.

[1] https://github.com/wenzhixin/bootstrap-table/commit/8a46a228