Disable --strict-optional in mypy
strict optional checks are not viable unless we have a solution for
the construction vs. representation mismatch of sqlalchemy ORM
objects:  For instance, although `User.login` is not nullable on the
database level, `User().login` is certainly None.
However, the non-nullability is guaranteed for any `user: User` object
which represents an _existing_ object in the database.
Both scenarios fall under the umbrella of the `User` type, yet a distinction
like `PartialUser` vs. `User` would be more appropriate to adequately
identify the nullability properties of said objects.

However, it's not clear if this can be solved in python/sqlalchemy at
the current point in time without introducing lots of boilerplate
and/or visual noise.

Refs #559