Do conversion to `Record` types in soure modules
This allows us to treat the user/group/property fetching as a one-step
process rather than doing two steps (fetching and then processing to
`Record`s later down the line), even if it makes things seem a little
more complicated at the moment.

Naming the methods `_records` is just a temporary thing to avoid a
name collision and not cause confusion during the refactoring.