[tests] Provide `test_client` in fixtures of different scope
This way one can easily override the `app` in a class when needed
without having to re-define a new, more tightly scoped `test_client`
fixture – or decide to leave everything as-is by using the
`module`-scoped fixture.

At the moment, there is no need to switch apps between different
functions in the same test class, so a `function` scoped `session`
object has been omitted.

In general, we want to use the coarsest scope anyway.

Refs #580