[docker] improve venv hook performance
1. Don't run `python -m ensurepip` every time

Instead we make the assumption that `ensurepip` has been run when the
virtualenv was created.

2. Don't run `pip freeze` every time

Instead we determine whether the pip packages have been installed by
observing whether `pycroft` can be imported.

This cuts off a few seconds on every `docker compose run --rm test-app
shell` invocation; in particular, it greatly reduces the overhead to
start a python shell or a test in PyCharm.