Move fints helper module to new `external_services` package
This is a good way to separate the components depending on external
services.

It is advisable that these services don't depend on the pycroft model,
but instead provide “data transfer objects” (like `StatementError`).

This allows for these modules to be tested separately, and restricts
what happens in there to translation / orchestration, while forcing
the more pycroft-specific transformation logic to happen in `lib`.