[lib] initialize logger where needed instead of in `lib.finance`
The current state broke improts, because subpackages tried to import
`lib.finance.logger`.
There would have been two other solutions:

1. Move the `logger` declaration to before the imports from the
   subpackages
2. Move the `logger` to a `lib.finance.logging` package, which then
   would just contain that logger.

Variant 1 seemed inconsistent with the usual import structure, while 2
looked like a waste of the file.

This way, everything works, but if a package-level logger is needed,
we can add it as required.