[model] use overlaps for `BankAccountActivity` relationships
Both a bank account activity and a Split reference an Account and a
Transaction.

Indeed, if a BankAccountActivity is matched to a transaction, it must
be matched to the (unique!) split of the transaction belonging to the
bank account.

This is an „overlap“ in foreign keys, whose consistency is checked by
a trigger.  The `overlaps` parameter is there to assure sqlalchemy we
know what we're doing.

Indeed, setting `transaction.splits.append(Split(…))` will override
the `transaction_id` of the split; this may override the
`transaction_id` set by `split.bank_account_activity =
BankAccountActivity(…)`. However, in any scenario, the overwrite would
produce the same value, since this is the consistency the trigger
guarantees.